server/requirements.txt
argcomplete @ file:///private/tmp/python-argcomplete-20240123-4975-yeh326/argcomplete-3.2.2
certifi @ file:///private/tmp/python-certifi-20240201-4765-6ufe0c/certifi-2024.2.2
dnspython==1.16.0
packaging @ file:///private/tmp/python-packaging-20231211-10905-adsfgp/packaging-23.2
pymongo==3.6.0
python-dotenv==1.0.1
zeroc-ice==3.7.10
